+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The current role maintainer is the SBB Cloud Platform Team.
 
 ### Added
 
+- Function `getJobTemplateDetails()` and API route `/tower/job_templates/<id>/getDetails` (GET)
+  added, to get survey specs from Ansible Tower templates.
 - The "functional account" (an additional project admin for Openshift) is now configurable and not
   hardcoded into the go file.
 - New tests for the validateProjectPermissions() function from openshift/project.go